+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/impl/parse/ensembl/EnsemblParser.java
@@ -1,10 +1,6 @@
 package de.charite.compbio.jannovar.impl.parse.ensembl;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
+import com.google.common.collect.*;
 import de.charite.compbio.jannovar.JannovarException;
 import de.charite.compbio.jannovar.UncheckedJannovarException;
 import de.charite.compbio.jannovar.data.ReferenceDictionary;
@@ -21,23 +17,15 @@ import de.charite.compbio.jannovar.reference.GenomeInterval;
 import de.charite.compbio.jannovar.reference.Strand;
 import de.charite.compbio.jannovar.reference.TranscriptModel;
 import de.charite.compbio.jannovar.reference.TranscriptModelBuilder;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.zip.GZIPInputStream;
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorInputStream;
 import org.ini4j.Profile.Section;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.io.*;
+import java.util.*;
+import java.util.Map.Entry;
+import java.util.zip.GZIPInputStream;
 
 // TODO(holtgrewe): Factor out common paths with RefSeqParser
 // TODO(holtgrewe): stop codon part of CDS here?
@@ -59,6 +47,10 @@ public class EnsemblParser implements TranscriptParser {
 	 * Path to the {@link ReferenceDictionary} to use for name/id and id/length mapping
 	 */
 	private final ReferenceDictionary refDict;
+	/*
+	 * Contig name mappings
+	 */
+	private final ImmutableMap<String, Integer> contigDict;
 
 	/**
 	 * Path to directory where the to-be-parsed files live
@@ -84,6 +76,7 @@ public class EnsemblParser implements TranscriptParser {
 	public EnsemblParser(ReferenceDictionary refDict, String basePath, List<String> geneIdentifiers,
 		Section iniSection) {
 		this.refDict = refDict;
+		this.contigDict = refDict.getContigNameToID();
 		this.basePath = basePath;
 		this.iniSection = iniSection;
 		this.geneIdentifiers = geneIdentifiers;
@@ -93,26 +86,18 @@ public class EnsemblParser implements TranscriptParser {
 	public ImmutableList<TranscriptModel> run() throws TranscriptParseException {
 		// Load features from GTF file, clustered by the gene they belong to
 		final String pathGTF = PathUtil.join(basePath, getINIFileName("gtf"));
-		Map<String, TranscriptModelBuilder> builders = recordsToBuilders(loadRecords(pathGTF));
+		Map<String, TranscriptModelBuilder> builders = loadTranscriptModels(pathGTF);
 
-		// Load files that allow mapping from ENSG to HGNC ID.
-		final Map<String, String> ensgToHgnc;
-		try {
-			ensgToHgnc = loadEngsToHgnc();
-		} catch (Exception e) {
-			throw new TranscriptParseException("Could not parse ENSEMBL mapping files", e);
-		}
+		// Load mappings that allows mapping from ENSG to HGNC ID.
+		LOGGER.info("Loading ENSEMBL to HGNC mappings...");
+		final Map<String, String> ensgToHgnc = loadEngsToHgnc();
 
 		// Load files that give the ENSG to Entrez ID mapping that comes directly from ENSEMBL.
-		final Map<String, String> ensgToEntrez;
-		try {
-			ensgToEntrez = loadEngsToEntrez();
-		} catch (Exception e) {
-			throw new TranscriptParseException(
-				"Could not parse ENSEMBL mapping files ENSG to Entrez", e);
-		}
+		LOGGER.info("Loading ENSEMBL to ENTREZ mappings...");
+		final Map<String, String> ensgToEntrez = loadEngsToEntrez();
 
 		// Augment information in builders with
+		LOGGER.info("Assigning additional HGNC information to transcripts..");
 		try {
 			new TranscriptModelBuilderHGNCExtender(
 				basePath,
@@ -126,9 +111,6 @@ public class EnsemblParser implements TranscriptParser {
 
 		// Use Entrez IDs from RefSeq if no HGNC annotation
 		for (TranscriptModelBuilder val : builders.values()) {
-			if ("ENSG00000272333".equals(val.getGeneID())) {
-				System.err.println("===>>>===>>> RARKENDARL");
-			}
 			if (val.getAltGeneIDs().isEmpty() && val.getGeneID() != null) {
 				if (ensgToEntrez.containsKey(val.getGeneID())) {
 					final String entrezGeneId = ensgToEntrez.get(val.getGeneID());
@@ -142,24 +124,22 @@ public class EnsemblParser implements TranscriptParser {
 					new Object[]{val.getGeneID(), val.getGeneID()});
 				val.getAltGeneIDs().put(AltGeneIDType.ENSEMBL_GENE_ID.toString(), val.getGeneID());
 			}
-			if ("ENSG00000272333".equals(val.getGeneID())) {
-				System.err.println("===<<<===<<< RARKENDARL");
-			}
 		}
 
 		// Load the FASTA file and assign to the builders.
 		final String pathFASTA = PathUtil.join(basePath, getINIFileName("cdna"));
+		LOGGER.info("Adding sequence information from cdna FASTA...");
 		loadFASTA(builders, pathFASTA);
 
+		LOGGER.info("Finalising TranscriptModels...");
 		// Create final list of TranscriptModels.
-		ImmutableList.Builder<TranscriptModel> result = new ImmutableList.Builder<TranscriptModel>();
+		ImmutableList.Builder<TranscriptModel> result = new ImmutableList.Builder<>();
 		for (Entry<String, TranscriptModelBuilder> entry : builders.entrySet()) {
 			TranscriptModelBuilder builder = entry.getValue();
 			if (geneIdentifiers == null || geneIdentifiers.isEmpty()) {
 				result.add(builder.build());
 			} else {
-				if (geneIdentifiers.contains(builder.getAccession()) || geneIdentifiers
-					.contains(builder.getGeneID())
+				if (geneIdentifiers.contains(builder.getAccession()) || geneIdentifiers.contains(builder.getGeneID())
 					|| !Sets.intersection(ImmutableSet.copyOf(geneIdentifiers),
 					ImmutableSet.copyOf(builder.getAltGeneIDs().values())).isEmpty()
 
@@ -168,10 +148,20 @@ public class EnsemblParser implements TranscriptParser {
 				}
 			}
 		}
-		return result.build();
+		ImmutableList<TranscriptModel> transcriptModels = result.build();
+		LOGGER.info("Built {} TranscriptModels", transcriptModels.size());
+		return transcriptModels;
 	}
 
-	private Map<String, String> loadEngsToEntrez() throws IOException {
+	/**
+	 * @param key name of the INI entry
+	 * @return file name from INI <code>key</code.
+	 */
+	private String getINIFileName(String key) {
+		return new File(iniSection.get(key)).getName();
+	}
+
+	private Map<String, String> loadEngsToEntrez() throws TranscriptParseException {
 		// Read mapping from ENSG to MySQL key
 		final String pathTableGeneMain = PathUtil.join(basePath, getINIFileName("table_gene_main"));
 		final Map<String, String> ensgToKey = new HashMap<>();
@@ -189,6 +179,8 @@ public class EnsemblParser implements TranscriptParser {
 				final String[] arr = line.trim().split("\t");
 				ensgToKey.put(arr[6], arr[1]);
 			}
+		} catch (Exception e) {
+			throw new TranscriptParseException("Could not parse ENSEMBL mapping files", e);
 		}
 
 		// Read mapping from MySQL key to Entrez ID
@@ -211,6 +203,8 @@ public class EnsemblParser implements TranscriptParser {
 					keyToEntrez.put(arr[0], arr[3]);
 				}
 			}
+		} catch (Exception e) {
+			throw new TranscriptParseException("Could not parse ENSEMBL mapping files", e);
 		}
 
 		// Build mapping from ENSG to Entrez ID
@@ -228,7 +222,8 @@ public class EnsemblParser implements TranscriptParser {
 		return result;
 	}
 
-	private Map<String, String> loadEngsToHgnc() throws IOException {
+
+	private Map<String, String> loadEngsToHgnc() throws TranscriptParseException {
 		// Read mapping from ENSG to MySQL key
 		final String pathTableGeneMain = PathUtil.join(basePath, getINIFileName("table_gene_main"));
 		final Map<String, String> ensgToKey = new HashMap<>();
@@ -246,6 +241,8 @@ public class EnsemblParser implements TranscriptParser {
 				final String[] arr = line.trim().split("\t");
 				ensgToKey.put(arr[6], arr[1]);
 			}
+		} catch (Exception e) {
+			throw new TranscriptParseException("Could not parse ENSEMBL mapping files", e);
 		}
 
 		// Read mapping from MySQL key to HGNC ID
@@ -267,6 +264,8 @@ public class EnsemblParser implements TranscriptParser {
 					keyToHgnc.put(arr[0], arr[3]);
 				}
 			}
+		} catch (Exception e) {
+			throw new TranscriptParseException("Could not parse ENSEMBL mapping files", e);
 		}
 
 		// Build mapping from ENSG to HGNC identifier
@@ -276,7 +275,7 @@ public class EnsemblParser implements TranscriptParser {
 			final String key = entry1.getValue();
 			final String hgnc = keyToHgnc.get(key);
 			if (hgnc == null) {
-				LOGGER.warn("Found no HGNC identifier for ENSG: ", new Object[]{ensg});
+				LOGGER.debug("Found no HGNC identifier for ENSG: {}", ensg);
 			} else {
 				result.put(ensg, hgnc);
 			}
@@ -291,19 +290,19 @@ public class EnsemblParser implements TranscriptParser {
 	 */
 	private void loadFASTA(Map<String, TranscriptModelBuilder> builders, String pathFASTA)
 		throws TranscriptParseException {
+
 		// First, build mapping from RNA accession to builder
 		Map<String, TranscriptModelBuilder> txMap = new HashMap<>();
 		for (Entry<String, TranscriptModelBuilder> entry : builders.entrySet()) {
-			txMap.put(entry.getValue().getSequence(), entry.getValue());
-			if (entry.getValue().getTxVersion() != null) {
-				txMap.put(entry.getValue().getSequence() + "." + entry.getValue().getTxVersion(),
-					entry.getValue());
+			TranscriptModelBuilder trnscpModelBuilder = entry.getValue();
+			txMap.put(trnscpModelBuilder.getSequence(), trnscpModelBuilder);
+			if (trnscpModelBuilder.getTxVersion() != null) {
+				txMap.put(trnscpModelBuilder.getSequence() + "." + trnscpModelBuilder.getTxVersion(), trnscpModelBuilder);
 			}
 		}
 
-		// We must remove variants for which we did not find any sequence;
-		Set<String> missingSequence = new HashSet<>();
-		missingSequence.addAll(builders.keySet());
+		// We must remove variants for which we did not find any sequence
+		Set<String> missingSequence = new HashSet<>(builders.keySet());
 
 		// Next iterate over the FASTA file and assign sequence to the transcript
 		FASTAParser fastaParser;
@@ -319,8 +318,7 @@ public class EnsemblParser implements TranscriptParser {
 				final TranscriptModelBuilder builder = txMap.get(accession);
 				if (builder == null) {
 					// This is not a warning as we observed this for some records regularly
-					LOGGER.debug("ID {} from FASTA did not map to transcript",
-						new Object[]{accession});
+					LOGGER.debug("ID {} from FASTA did not map to transcript", accession);
 					continue;
 				}
 
@@ -329,22 +327,19 @@ public class EnsemblParser implements TranscriptParser {
 
 				builder.setAccession(builder.getSequence());
 				builder.setSequence(record.getSequence());
-				LOGGER.debug("Found sequence for transcript {}",
-					new Object[]{builder.getAccession()});
+				LOGGER.debug("Found sequence for transcript {}", builder.getAccession());
 			}
 		} catch (IOException e) {
 			throw new TranscriptParseException("Problem with reading FASTA file", e);
 		}
 
-		LOGGER.info("Ignoring {} transcripts without sequence.",
-			new Object[]{missingSequence.size()});
+		LOGGER.info("Ignoring {} transcripts without sequence.", missingSequence.size());
 		for (String key : missingSequence) {
-			LOGGER.debug("--> {}", new Object[]{key});
+			LOGGER.debug("--> {}", key);
 			builders.remove(key);
 		}
 
-		LOGGER.info("Successfully processed {} transcripts with sequence.",
-			new Object[]{builders.size()});
+		LOGGER.info("Successfully processed {} transcripts with sequence.", builders.size());
 	}
 
 	/**
@@ -356,128 +351,10 @@ public class EnsemblParser implements TranscriptParser {
 	 * The TranscriptModelBuilder objects will have the a "Name" attribute of the mRNA set as the
 	 * sequence, so we can use this for assigning FASTA sequence to the builders.
 	 */
-	private Map<String, TranscriptModelBuilder> recordsToBuilders(
-		HashMap<String, ArrayList<FeatureRecord>> recordsByGene) {
-		Map<String, TranscriptModelBuilder> result = new HashMap<>();
-		for (Entry<String, ArrayList<FeatureRecord>> entry : recordsByGene.entrySet()) {
-			result.putAll(processGeneGFFRecords(entry.getValue()));
-		}
-		return result;
-	}
-
-	/**
-	 * Process the GFFRecord objects for one gene.
-	 */
-	private Map<String, TranscriptModelBuilder> processGeneGFFRecords(
-		ArrayList<FeatureRecord> records) {
-		final Map<String, TranscriptModelBuilder> result = new HashMap<>();
-
-		// Factorize the records by the transcript ID
-		final HashMap<String, ArrayList<FeatureRecord>> recordsForTX = new HashMap<>();
-		for (FeatureRecord record : records) {
-			final String txID = (record.getAttributes().get("transcript_id") != null)
-				? record.getAttributes().get("transcript_id")
-				: record.getAttributes().get("transcript_name");
-			if (txID == null) {
-				continue; // skip, no transcript ID
-			}
-			if (!recordsForTX.containsKey(txID)) {
-				recordsForTX.put(txID, Lists.newArrayList(record));
-			} else {
-				recordsForTX.get(txID).add(record);
-			}
-		}
-
-		// Now, build TranscriptModelBuilder for each transcript
-		for (Entry<String, ArrayList<FeatureRecord>> txEntry : recordsForTX.entrySet()) {
-			final List<FeatureRecord> featureRecords = txEntry.getValue();
-
-			final FeatureRecord first = featureRecords.get(0);
-			final String geneName = first.getAttributes().get("gene_name");
-			final String geneID = first.getAttributes().get("gene_id");
-			final String txID = first.getAttributes().get("transcript_id");
-			final String txVersion = first.getAttributes().get("transcript_version");
-
-			final TranscriptModelBuilder builder = new TranscriptModelBuilder();
-
-			// Parse out the simple attributes from the mRNA record
-			final Strand strand =
-				(first.getStrand() == FeatureRecord.Strand.FORWARD) ? Strand.FWD : Strand.REV;
-			builder.setStrand(strand);
-			builder.setAccession(txID);
-			builder.setTxVersion(txVersion);
-			builder.setGeneID(geneID);
-			builder.setGeneSymbol(geneName);
-			builder.setSequence(txID);
-
-			// Iterate over the features, interpreting "exon" and "CDS"/"stop_codon" entries
-			GenomeInterval txRegion = null;
-			GenomeInterval cdsRegion = null;
-			boolean wrongContig = false;
-			for (FeatureRecord record : featureRecords) {
-				ImmutableMap<String, Integer> dict = refDict.getContigNameToID();
-				final String seqID = record.getSeqID();
-				if (!dict.containsKey(seqID)) {
-					LOGGER.debug("Skipping record {} on unknown contig {}",
-						new Object[]{record, seqID});
-					wrongContig = true;
-					continue;
-				}
-				if (record.getType().equals("exon")) {
-					final int chrom = dict.get(seqID);
-					GenomeInterval exon = new GenomeInterval(refDict, Strand.FWD, chrom,
-						record.getBegin(),
-						record.getEnd());
-					exon = exon.withStrand(strand);
-					if (txRegion == null) {
-						txRegion = exon;
-					} else {
-						txRegion = txRegion.union(exon);
-					}
-					builder.addExonRegion(exon);
-				} else if ("CDS".equals(record.getType()) || "stop_codon"
-					.equals(record.getType())) {
-					GenomeInterval cds = new GenomeInterval(refDict, Strand.FWD,
-						refDict.getContigNameToID().get(record.getSeqID()), record.getBegin(),
-						record.getEnd());
-					cds = cds.withStrand(strand);
-					if (cdsRegion == null) {
-						cdsRegion = cds;
-					} else {
-						cdsRegion = cdsRegion.union(cds);
-					}
-				}
-			}
-			if (wrongContig) {
-				continue; // skip, on wrong contig
-			}
-			if (txRegion == null) {
-				// Only warn if a transcript and not a gene, we only allow exons to be parts of genes as this is
-				// observed in RefSeq
-				LOGGER.error("No transcript region for {}; skipping", new Object[]{txEntry});
-				continue;
-			}
-			builder.setTXRegion(txRegion);
-			if (cdsRegion == null) {
-				cdsRegion = new GenomeInterval(txRegion.getGenomeBeginPos(), 0);
-			}
-			builder.setCDSRegion(cdsRegion);
-
-			result.put(txID, builder);
-		}
-
-		return result;
-	}
-
-	/**
-	 * Load GFF records, cluster by gene and return
-	 *
-	 * @throws TranscriptParseException on problems with handling the transcript file
-	 */
-	private HashMap<String, ArrayList<FeatureRecord>> loadRecords(String pathGFF)
-		throws TranscriptParseException {
-		HashMap<String, ArrayList<FeatureRecord>> result = new HashMap<String, ArrayList<FeatureRecord>>();
-
+	private Map<String, TranscriptModelBuilder> loadTranscriptModels(String pathGFF) throws TranscriptParseException {
+		LOGGER.info("Loading feature records");
+		// transcriptId: TranscriptModelBuilder
+		Map<String, TranscriptModelBuilder> results = new HashMap<>(200000);
 		// Open file using GFFParser
 		GFFParser parser;
 		try {
@@ -486,6 +363,7 @@ public class EnsemblParser implements TranscriptParser {
 			throw new TranscriptParseException("Problem opening GFF file", e);
 		}
 
+		Set<String> wantedTypes = Sets.newHashSet("exon", "CDS", "stop_codon");
 		// Read file record by record, mapping features to genes
 		//
 		// This will only work properly if the full path of feature objects from the current feature has already been
@@ -494,31 +372,98 @@ public class EnsemblParser implements TranscriptParser {
 		try {
 			FeatureRecord record;
 			while ((record = parser.next()) != null) {
-				LOGGER.debug("Loaded GFF record {}", new Object[]{record});
-				numRecords += 1;
-
-				final String geneID = record.getAttributes().get("gene_id");
-				if (!result.containsKey(geneID)) {
-					result.put(geneID, Lists.newArrayList(record));
-				} else {
-					result.get(geneID).add(record);
+				numRecords++;
+				// filter these out here as they are only discarded later
+				if (record.getAttributes()
+					.get("transcript_id") != null && contigDict.containsKey(record.getSeqID()) && wantedTypes.contains(record
+					.getType())) {
+					LOGGER.debug("Loaded GFF record {}", record);
+					String transcriptId = record.getAttributes().get("transcript_id");
+					if (!results.containsKey(transcriptId)) {
+						// create new TranscriptBuilder
+						TranscriptModelBuilder builder = createNewTranscriptModelBuilder(record, transcriptId);
+						results.put(transcriptId, builder);
+					} else {
+						// update existing
+						TranscriptModelBuilder builder = results.get(transcriptId);
+						updateExonsTxRegionsAndCds(record, builder);
+					}
 				}
+
 			}
 		} catch (IOException e) {
 			throw new TranscriptParseException("Problem parsing GFF file", e);
 		}
 
-		LOGGER.info("Loaded {} GFF records for {} genes", new Object[]{numRecords, result.size()});
+		Map<String, TranscriptModelBuilder> transcriptModelsWithTxRegion = getTranscriptModelsWithTxRegion(results);
+		LOGGER.info("Parsed {} GFF records as {} TranscriptModels", numRecords, transcriptModelsWithTxRegion.size());
 
-		return result;
+		return transcriptModelsWithTxRegion;
 	}
 
-	/**
-	 * @param key name of the INI entry
-	 * @return file name from INI <code>key</code.
-	 */
-	private String getINIFileName(String key) {
-		return new File(iniSection.get(key)).getName();
+	private TranscriptModelBuilder createNewTranscriptModelBuilder(FeatureRecord record, String transcriptId) {
+		TranscriptModelBuilder builder = new TranscriptModelBuilder();
+		// Parse out the simple attributes from the mRNA record
+		Strand strand = parseStrand(record);
+		builder.setStrand(strand);
+		builder.setAccession(transcriptId);
+		builder.setTxVersion(record.getAttributes().get("transcript_version"));
+		builder.setGeneID(record.getAttributes().get("gene_id"));
+		builder.setGeneSymbol(record.getAttributes().get("gene_name"));
+		builder.setSequence(transcriptId);
+
+		updateExonsTxRegionsAndCds(record, builder);
+
+		return builder;
+	}
+
+	private void updateExonsTxRegionsAndCds(FeatureRecord record, TranscriptModelBuilder builder) {
+		Strand strand = parseStrand(record);
+		if (record.getType().equals("exon")) {
+			GenomeInterval exon = buildGenomeInterval(record, strand);
+			GenomeInterval txRegion = updateGenomeInterval(exon, builder.getTXRegion());
+			builder.setTXRegion(txRegion);
+			builder.addExonRegion(exon);
+		} else if ("CDS".equals(record.getType()) || "stop_codon".equals(record.getType())) {
+			GenomeInterval cds = buildGenomeInterval(record, strand);
+			GenomeInterval cdsRegion = updateGenomeInterval(cds, builder.getCDSRegion());
+			builder.setCDSRegion(cdsRegion);
+		}
+	}
+
+	private GenomeInterval updateGenomeInterval(GenomeInterval latest, GenomeInterval existing) {
+		return existing == null ? latest : existing.union(latest);
+	}
+
+	private Strand parseStrand(FeatureRecord record) {
+		return (record.getStrand() == FeatureRecord.Strand.FORWARD) ? Strand.FWD : Strand.REV;
+	}
+
+	private GenomeInterval buildGenomeInterval(FeatureRecord record, Strand strand) {
+		int chrom = contigDict.get(record.getSeqID());
+		GenomeInterval interval = new GenomeInterval(refDict, Strand.FWD, chrom, record.getBegin(), record.getEnd());
+		// CAUTION! GFF record begin and end are listed using the FORWARD strand, so this needs adjusting-post build
+		// rather than being supplied in the constructor.
+		return interval.withStrand(strand);
+	}
+
+	private Map<String, TranscriptModelBuilder> getTranscriptModelsWithTxRegion(Map<String, TranscriptModelBuilder> results) {
+		Map<String, TranscriptModelBuilder> transcriptModelsWithTxRegion = new HashMap<>(results.size());
+
+		results.forEach((transcriptId, transcriptModelBuilder) -> {
+			GenomeInterval txRegion = transcriptModelBuilder.getTXRegion();
+			if (txRegion == null) {
+				// Only warn if a transcript and not a gene, we only allow exons to be parts of genes as this is
+				// observed in RefSeq
+				LOGGER.warn("No transcript region for {}; skipping", transcriptId);
+			} else {
+				if (transcriptModelBuilder.getCDSRegion() == null) {
+					transcriptModelBuilder.setCDSRegion(new GenomeInterval(txRegion.getGenomeBeginPos(), 0));
+				}
+				transcriptModelsWithTxRegion.put(transcriptId, transcriptModelBuilder);
+			}
+		});
+		return transcriptModelsWithTxRegion;
 	}
 
 }