+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/impl/parse/refseq/RefSeqParser.java
@@ -22,11 +22,13 @@ import org.ini4j.Profile.Section;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.annotation.Nullable;
 import java.io.File;
 import java.io.IOException;
 import java.util.*;
 import java.util.Map.Entry;
 
+import static java.util.stream.Collectors.*;
 /**
  * Parsing of RefSeq GFF3 files
  *
@@ -49,6 +51,11 @@ public class RefSeqParser implements TranscriptParser {
 	 * Path to the {@link ReferenceDictionary} to use for name/id and id/length mapping
 	 */
 	private ReferenceDictionary refDict;
+
+	/*
+	 * Contig name mappings
+	 */
+	private final ImmutableMap<String, Integer> contigDict;
 
 	/**
 	 * Path to directory where the to-be-parsed files live
@@ -74,6 +81,8 @@ public class RefSeqParser implements TranscriptParser {
 	public RefSeqParser(ReferenceDictionary refDict, String basePath, List<String> geneIdentifiers,
 						Section iniSection) {
 		this.refDict = refDict;
+		this.contigDict = refDict.getContigNameToID();
+
 		this.basePath = basePath;
 		this.iniSection = iniSection;
 		this.geneIdentifiers = geneIdentifiers;
@@ -83,12 +92,13 @@ public class RefSeqParser implements TranscriptParser {
 	public ImmutableList<TranscriptModel> run() throws TranscriptParseException {
 		// Load features from GFF3 file, clustered by the gene they belong to
 		final String pathGFF = PathUtil.join(basePath, getINIFileName("gff"));
-		Map<String, TranscriptModelBuilder> builders = recordsToBuilders(loadRecords(pathGFF));
+		Map<String, TranscriptModelBuilder> builders = loadTranscriptModels(pathGFF);
 
-		// Augment information in builders with
+		// Augment information in builders with HGNC mappings
+		LOGGER.info("Assigning additional HGNC information to {} transcripts..", builders.size());
 		try {
 			new TranscriptModelBuilderHGNCExtender(basePath, r -> Lists.newArrayList(r.getEntrezID()),
-				tx -> tx.getGeneID()).run(builders);
+				TranscriptModelBuilder::getGeneID).run(builders);
 		} catch (JannovarException e) {
 			throw new UncheckedJannovarException("Problem extending transcripts with HGNC information", e);
 		}
@@ -96,15 +106,15 @@ public class RefSeqParser implements TranscriptParser {
 		// Use Entrez IDs from RefSeq if no HGNC annotation
 		for (TranscriptModelBuilder val : builders.values()) {
 			if (val.getAltGeneIDs().isEmpty() && val.getGeneID() != null) {
-				LOGGER.info("Using UCSC Entrez ID {} for transcript {} as HGNC did not provide alternative gene ID",
-					new Object[]{val.getGeneID(), val.getAccession()});
+				LOGGER.debug("Using UCSC Entrez ID {} for transcript {} as HGNC did not provide alternative gene ID",
+					val.getGeneID(), val.getAccession());
 				val.getAltGeneIDs().put(AltGeneIDType.ENTREZ_ID.toString(), val.getGeneID());
 			}
 		}
 
 		// Load the FASTA file and assign to the builders.
 		final String pathFASTA = PathUtil.join(basePath, getINIFileName("rna"));
-		loadFASTA(builders, pathFASTA, preferPARTranscriptsOnChrX());
+		loadFASTA(builders, pathFASTA);
 
 		// Create final list of TranscriptModels.
 		ImmutableList.Builder<TranscriptModel> result = new ImmutableList.Builder<>();
@@ -126,18 +136,22 @@ public class RefSeqParser implements TranscriptParser {
 	}
 
 	/**
+	 * @param key name of the INI entry
+	 * @return file name from INI <code>key</code.
+	 */
+	private String getINIFileName(String key) {
+		return new File(iniSection.get(key)).getName();
+	}
+
+	/**
 	 * Load FASTA from pathFASTA and set the sequence into builders.
 	 *
 	 * @throws TranscriptParseException on problems with parsing the FASTA
 	 */
-	private void loadFASTA(Map<String, TranscriptModelBuilder> builders, String pathFASTA,
-						   boolean preferPARTranscriptsOnChrX) throws TranscriptParseException {
-		// First, build mapping from RNA accession to builder
-		Map<String, TranscriptModelBuilder> txMap = mapFromAccessionToBuilder(builders, preferPARTranscriptsOnChrX);
+	private void loadFASTA(Map<String, TranscriptModelBuilder> builders, String pathFASTA) throws TranscriptParseException {
 
-		// We must remove variants for which we did not find any sequence;
-		Set<String> missingSequence = new HashSet<>();
-		missingSequence.addAll(builders.keySet());
+		// We must remove variants for which we did not find any sequence
+		Set<String> missingSequence = new HashSet<>(builders.keySet());
 
 		// Next iterate over the FASTA file and assign sequence to the transcript
 		FASTAParser fastaParser;
@@ -154,58 +168,31 @@ public class RefSeqParser implements TranscriptParser {
 					continue;
 				}
 				String accession = accessionOpt.get();
-				final TranscriptModelBuilder builder = txMap.get(accession);
+				final TranscriptModelBuilder builder = builders.get(accession);
 				if (builder == null) {
 					// This is not a warning as we observed this for some records regularly
-					LOGGER.debug("ID {} from FASTA did not map to transcript", new Object[]{accession});
+					LOGGER.debug("ID {} from FASTA did not map to transcript", accession);
 					continue;
 				}
 
-				assert missingSequence.contains(builder.getAccession());
-				missingSequence.remove(builder.getAccession());
+				assert missingSequence.contains(accession);
+				missingSequence.remove(accession);
 
 				builder.setAccession(accession);
 				builder.setSequence(record.getSequence());
-				LOGGER.debug("Found sequence for transcript {}", new Object[]{builder.getAccession()});
+				LOGGER.debug("Found sequence for transcript {}", builder.getAccession());
 			}
 		} catch (IOException e) {
 			throw new TranscriptParseException("Problem with reading FASTA file", e);
 		}
 
-		LOGGER.info("Ignoring {} transcripts without sequence.", new Object[]{missingSequence.size()});
+		LOGGER.info("Ignoring {} transcripts without sequence.", missingSequence.size());
 		for (String key : missingSequence) {
-			LOGGER.debug("--> {}", new Object[]{key});
+			LOGGER.debug("--> {}", key);
 			builders.remove(key);
 		}
 
-		LOGGER.info("Successfully processed {} transcripts with sequence.", new Object[]{builders.size()});
-	}
-
-	private static Map<String, TranscriptModelBuilder> mapFromAccessionToBuilder(Map<String, TranscriptModelBuilder> builders,
-																				 boolean preferPARTranscriptsOnChrX) {
-		Map<String, TranscriptModelBuilder> txMap = new HashMap<>();
-		for (Entry<String, TranscriptModelBuilder> entry : builders.entrySet()) {
-			if (entry.getValue().getSequence() == null) {
-				continue;
-			}
-			TranscriptModelBuilder existingTranscriptBuilder = txMap.get(entry.getValue().getSequence());
-			if (existingTranscriptBuilder != null) {
-				int newChromosome = entry.getValue().getTXRegion().getChr();
-				int currentChromosome = existingTranscriptBuilder.getTXRegion().getChr();
-				if (preferPARTranscriptsOnChrX && currentChromosome == 23 && newChromosome == 24) {
-					LOGGER.warn(
-						"PAR transcripts on chrX are configured to be preferable, so ignoring entry for chromosome '{}' in favor of entry for chromosome '{}' for '{}'",
-						newChromosome, currentChromosome, entry.getValue().getSequence());
-					continue;
-				} else {
-					LOGGER.warn(
-						"Found two locations for '{}', now using entry for chromosome '{}' and discarding the entry for chromosome '{}'",
-						entry.getValue().getSequence(), newChromosome, currentChromosome);
-				}
-			}
-			txMap.put(entry.getValue().getSequence(), entry.getValue());
-		}
-		return txMap;
+		LOGGER.info("Successfully processed {} transcripts with sequence.", builders.size());
 	}
 
 	/**
@@ -217,144 +204,9 @@ public class RefSeqParser implements TranscriptParser {
 	 * The TranscriptModelBuilder objects will have the a "Name" attribute of the mRNA set as the sequence, so we can
 	 * use this for assigning FASTA sequence to the builders.
 	 */
-	private Map<String, TranscriptModelBuilder> recordsToBuilders(Map<String, List<FeatureRecord>> recordsByGene) {
-		Map<String, TranscriptModelBuilder> result = new HashMap<>();
-		for (Entry<String, List<FeatureRecord>> entry : recordsByGene.entrySet())
-			result.putAll(processGeneGFFRecords(entry.getValue()));
-		return result;
-	}
-
-	/**
-	 * Process the GFFRecord objects for one gene.
-	 */
-	private Map<String, TranscriptModelBuilder> processGeneGFFRecords(List<FeatureRecord> records) {
-		final Map<String, TranscriptModelBuilder> result = new HashMap<>();
-
-		assert records.get(0).getType().equals("gene");
-		final FeatureRecord geneRecord = records.get(0);
-
-		// Get IDs of the transcripts and create a list of GFFRecord objects for each
-		final Map<String, FeatureRecord> mrnaRecords = new HashMap<>();
-		final Map<String, List<FeatureRecord>> recordsForMRNA = new HashMap<>();
-		for (FeatureRecord record : records) {
-			// We will later assign exons and CDS features to tx level features, but exons can also be part of a gene
-			// only
-			if (TX_LEVEL_FEATURE_TYPES.contains(record.getType()) || "gene".equals(record.getType())) {
-				mrnaRecords.put(record.getAttributes().get("ID"), record);
-				recordsForMRNA.put(record.getAttributes().get("ID"), new ArrayList<>());
-			}
-		}
-
-		// Factorize the GFFRecords by mRNA
-		for (FeatureRecord record : records) {
-			if (record.getType().equals("exon") || record.getType().equals("CDS")) {
-				for (String parent : Splitter.on(',').split(record.getAttributes().get("Parent"))) {
-					if (recordsForMRNA.get(parent) != null)
-						recordsForMRNA.get(parent).add(record);
-				}
-			}
-		}
-
-		// Now, build TranscriptModelBuilder for each mRNA
-		for (Entry<String, FeatureRecord> mrnaEntry : mrnaRecords.entrySet()) {
-			final String mrnaID = mrnaEntry.getKey();
-			final FeatureRecord mrnaRecord = mrnaEntry.getValue();
-			final List<FeatureRecord> featureRecords = recordsForMRNA.get(mrnaID);
-
-			final TranscriptModelBuilder builder = new TranscriptModelBuilder();
-
-			// Parse out the simple attributes from the mRNA record
-			final Strand strand = (mrnaRecord.getStrand() == FeatureRecord.Strand.FORWARD) ? Strand.FWD : Strand.REV;
-			builder.setStrand(strand);
-			builder.setAccession(mrnaID);
-			builder.setGeneSymbol(geneRecord.getAttributes().get("Name"));
-			final String mrnaName = mrnaRecord.getAttributes().get("Name");
-			builder.setSequence(mrnaName);
-			parseGeneID(builder, geneRecord);
-
-			// Iterate over the features, interpreting "exon" and "CDS" entries
-			GenomeInterval txRegion = null;
-			GenomeInterval cdsRegion = null;
-			boolean wrongContig = false;
-			for (FeatureRecord record : featureRecords) {
-				ImmutableMap<String, Integer> dict = refDict.getContigNameToID();
-				final String seqID = record.getSeqID();
-				if (!dict.containsKey(seqID)) {
-					LOGGER.debug("Skipping record {} on unknown contig {}", new Object[]{record, seqID});
-					wrongContig = true;
-					continue;
-				}
-				if (record.getType().equals("exon")) {
-					final int chrom = dict.get(seqID);
-					GenomeInterval exon = new GenomeInterval(refDict, Strand.FWD, chrom, record.getBegin(),
-						record.getEnd());
-					exon = exon.withStrand(strand);
-					if (txRegion == null)
-						txRegion = exon;
-					else
-						txRegion = txRegion.union(exon);
-					builder.addExonRegion(exon);
-				} else if (record.getType().equals("CDS")) {
-					GenomeInterval cds = new GenomeInterval(refDict, Strand.FWD,
-						refDict.getContigNameToID().get(record.getSeqID()), record.getBegin(), record.getEnd());
-					cds = cds.withStrand(strand);
-					if (cdsRegion == null)
-						cdsRegion = cds;
-					else
-						cdsRegion = cdsRegion.union(cds);
-				}
-			}
-			if (wrongContig)
-				continue; // skip, on wrong contig
-			if (txRegion == null) {
-				// Only warn if a transcript and not a gene, we only allow exons to be parts of genes as this is
-				// observed in RefSeq
-				if (!mrnaRecord.getType().equals("gene"))
-					LOGGER.error("No transcript region for {}; skipping", new Object[]{mrnaEntry});
-				continue;
-			}
-			builder.setTXRegion(txRegion);
-			if (cdsRegion == null)
-				cdsRegion = new GenomeInterval(txRegion.getGenomeBeginPos(), 0);
-			builder.setCDSRegion(cdsRegion);
-
-			if (onlyCurated() && (mrnaName == null || mrnaName.startsWith("X"))) {
-				LOGGER.debug("Skipping non-curated transcript {}", new Object[]{mrnaName});
-				continue; // skip non-curated one
-			}
-
-			result.put(mrnaID, builder);
-		}
-
-		return result;
-	}
-
-	/**
-	 * Parse out entrez gene ID
-	 *
-	 * @param builder    The {@link TranscriptModelBuilder} to put the alternative geneIDs to
-	 * @param geneRecord {@link FeatureRecord} with the gene information
-	 */
-	private void parseGeneID(TranscriptModelBuilder builder, FeatureRecord geneRecord) {
-		if (!geneRecord.getAttributes().containsKey("Dbxref"))
-			return;
-		for (String token : Splitter.on(',').split(geneRecord.getAttributes().get("Dbxref"))) {
-			List<String> keyValue = Splitter.on(':').limit(2).splitToList(token);
-			if (keyValue.size() != 2)
-				continue;
-			if (keyValue.get(0).equals("GeneID"))
-				builder.setGeneID(keyValue.get(1));
-		}
-	}
-
-	/**
-	 * Load GFF records, cluster by gene and return
-	 *
-	 * @throws TranscriptParseException on problems with handling the transcript file
-	 */
-	private Map<String, List<FeatureRecord>> loadRecords(String pathGFF) throws TranscriptParseException {
-		Map<String, List<FeatureRecord>> result = new HashMap<>();
-
+	private Map<String, TranscriptModelBuilder> loadTranscriptModels(String pathGFF) throws TranscriptParseException {
+		LOGGER.info("Loading feature records");
+		// transcriptId: TranscriptModelBuilder
 		// Open file using GFFParser
 		GFFParser parser;
 		try {
@@ -362,48 +214,51 @@ public class RefSeqParser implements TranscriptParser {
 		} catch (IOException e) {
 			throw new TranscriptParseException("Problem opening GFF file", e);
 		}
+		// cached values from parsed records
+		Map<String, TranscriptModelBuilder> parentIdToTranscriptModels = new HashMap<>(200000);
+		Map<String, List<String>> transcriptIdToParentIds = new HashMap<>(parentIdToTranscriptModels.size());
 
-		// Map a feature to its gene
-		Map<String, String> featureToGene = new HashMap<>();
-
+		Set<String> wantedTypes = Sets.newHashSet("exon", "CDS", "stop_codon");
+		boolean onlyCurated = onlyCurated();
 		// Read file record by record, mapping features to genes
-		//
-		// This will only work properly if the full path of feature objects from the current feature has already been
-		// read. Otherwise, we will need some more fancy parsing.
 		int numRecords = 0;
 		try {
 			FeatureRecord record;
 			while ((record = parser.next()) != null) {
-				LOGGER.debug("Loaded GFF record {}", new Object[]{record});
-				final String id = record.getAttributes().get("ID");
-				numRecords += 1;
-				if ("gene".equals(record.getType())) {
-					LOGGER.debug("-> new gene {}", new Object[]{id});
-					featureToGene.put(id, id); // register mapping
-					assert !result.containsKey(id);
-					result.put(id, Lists.newArrayList(record));
-				} else {
-					final String parent = record.getAttributes().get("Parent");
-					if (parent == null)
-						continue; // ignore
-					final String top = featureToGene.get(parent);
-					if (top == null)
-						continue; // ignore, no gene entry
-					LOGGER.debug("-> parent = {}", new Object[]{parent});
-					LOGGER.debug("-> top = {}", new Object[]{top});
-					featureToGene.put(id, top); // register mapping
-					assert featureToGene.get(top).equals(top);
-					assert result.containsKey(top);
-					result.get(top).add(record);
+				numRecords++;
+				String transcriptId = record.getAttributes().get("transcript_id");
+				if (onlyCurated && (transcriptId == null || transcriptId.startsWith("X"))) {
+					LOGGER.debug("Skipping non-curated transcript {}", transcriptId);
+					continue;
+				}
+				// n.b. - in the RefSeq data the exon and CDS lines are linked by the Parent id as
+				// the transcriptId is only stated for the exon and the proteinId is used as the CDS identifier.
+				// a further complication is that some transcriptIds are used twice by different ParentIds - those in
+				// the pseudoautosomal regions and some others
+				String parentId = record.getAttributes().get("Parent");
+				if (parentId != null && contigDict.containsKey(record.getSeqID()) && wantedTypes.contains(record.getType())) {
+					if (!parentIdToTranscriptModels.containsKey(parentId)) {
+						// create new TranscriptBuilder
+						TranscriptModelBuilder builder = createNewTranscriptModelBuilder(record, transcriptId);
+						parentIdToTranscriptModels.put(parentId, builder);
+						// this method is used to track the rare cases of transcriptIds mapping to multiple parentIds
+						updateTransciptIdToParentIds(transcriptIdToParentIds, transcriptId, parentId);
+					} else {
+						// update existing
+						TranscriptModelBuilder builder = parentIdToTranscriptModels.get(parentId);
+						updateExonsTxRegionsAndCds(record, builder);
+					}
 				}
 			}
 		} catch (IOException e) {
 			throw new TranscriptParseException("Problem parsing GFF file", e);
 		}
 
-		LOGGER.info("Loaded {} GFF records for {} genes", new Object[]{numRecords, result.size()});
+		Map<String, TranscriptModelBuilder> transcriptIdToTranscriptModelBuilders = mapTranscriptIdsToTranscriptModels(parentIdToTranscriptModels, transcriptIdToParentIds);
 
-		return result;
+		LOGGER.info("Parsed {} GFF records as {} TranscriptModels", numRecords, transcriptIdToTranscriptModelBuilders.size());
+
+		return transcriptIdToTranscriptModelBuilders;
 	}
 
 	/**
@@ -431,12 +286,165 @@ public class RefSeqParser implements TranscriptParser {
 		return false;
 	}
 
-	/**
-	 * @param key name of the INI entry
-	 * @return file name from INI <code>key</code.
-	 */
-	private String getINIFileName(String key) {
-		return new File(iniSection.get(key)).getName();
+	private void updateTransciptIdToParentIds(Map<String, List<String>> transciptIdToParentIds, String transcriptId, String parentId) {
+		if (transcriptId != null) {
+			if (transciptIdToParentIds.containsKey(transcriptId)) {
+				List<String> parentIds = transciptIdToParentIds.get(transcriptId);
+				if (!parentIds.contains(parentId)) {
+					LOGGER.debug("Adding new parentId {} for transcript {}", parentId, transcriptId);
+					parentIds.add(parentId);
+				}
+			} else {
+				transciptIdToParentIds.put(transcriptId, Lists.newArrayList(parentId));
+			}
+		}
 	}
 
+	private TranscriptModelBuilder createNewTranscriptModelBuilder(FeatureRecord record, String transcriptId) {
+		TranscriptModelBuilder builder = new TranscriptModelBuilder();
+		// Parse out the simple attributes from the mRNA record
+		Strand strand = parseStrand(record);
+		builder.setStrand(strand);
+		builder.setAccession(transcriptId);
+		builder.setTxVersion(record.getAttributes().get("transcript_version"));
+		builder.setGeneID(parseGeneID(record));
+		builder.setGeneSymbol(record.getAttributes().get("gene"));
+		builder.setSequence(transcriptId);
+
+		updateExonsTxRegionsAndCds(record, builder);
+
+		return builder;
+	}
+
+	@Nullable
+	private String parseGeneID(FeatureRecord featureRecord) {
+		// Dbxref=GeneID:6010,HGNC:10012,HPRD:01584,MIM:180380;
+		String dbxrefs = featureRecord.getAttributes().get("Dbxref");
+		for (String token : Splitter.on(',').split(dbxrefs)) {
+			List<String> keyValue = Splitter.on(':').limit(2).splitToList(token);
+			if (keyValue.size() == 2 && keyValue.get(0).equals("GeneID")) {
+				return keyValue.get(1);
+			}
+		}
+
+		return null;
+	}
+
+	private void updateExonsTxRegionsAndCds(FeatureRecord record, TranscriptModelBuilder builder) {
+		Strand strand = parseStrand(record);
+		if (record.getType().equals("exon")) {
+			GenomeInterval exon = buildGenomeInterval(record, strand);
+			GenomeInterval txRegion = updateGenomeInterval(exon, builder.getTXRegion());
+			builder.setTXRegion(txRegion);
+			builder.addExonRegion(exon);
+		} else if ("CDS".equals(record.getType()) || "stop_codon".equals(record.getType())) {
+			GenomeInterval cds = buildGenomeInterval(record, strand);
+			GenomeInterval cdsRegion = updateGenomeInterval(cds, builder.getCDSRegion());
+			builder.setCDSRegion(cdsRegion);
+		}
+	}
+
+	private GenomeInterval updateGenomeInterval(GenomeInterval latest, GenomeInterval existing) {
+		return existing == null ? latest : existing.union(latest);
+	}
+
+	private Strand parseStrand(FeatureRecord record) {
+		return (record.getStrand() == FeatureRecord.Strand.FORWARD) ? Strand.FWD : Strand.REV;
+	}
+
+	private GenomeInterval buildGenomeInterval(FeatureRecord record, Strand strand) {
+		int chrom = contigDict.get(record.getSeqID());
+		GenomeInterval interval = new GenomeInterval(refDict, Strand.FWD, chrom, record.getBegin(), record.getEnd());
+		// CAUTION! GFF record begin and end are listed using the FORWARD strand, so this needs adjusting-post build
+		// rather than being supplied in the constructor.
+		return interval.withStrand(strand);
+	}
+
+	private Map<String, TranscriptModelBuilder> mapTranscriptIdsToTranscriptModels(Map<String, TranscriptModelBuilder> parentIdToTranscriptModels, Map<String, List<String>> transciptIdToParentIds) {
+		Map<String, TranscriptModelBuilder> parentIdToTranscriptModelsWithTxRegion = mapParentIdsToTranscriptModelsWithTxRegion(parentIdToTranscriptModels);
+		return mapNonRedundantTranscriptIdsToTranscriptModelBuilder(parentIdToTranscriptModelsWithTxRegion, transciptIdToParentIds);
+	}
+
+	private Map<String, TranscriptModelBuilder> mapParentIdsToTranscriptModelsWithTxRegion(Map<String, TranscriptModelBuilder> results) {
+		Map<String, TranscriptModelBuilder> transcriptModelsWithTxRegion = new HashMap<>(results.size());
+
+		results.forEach((parentId, transcriptModelBuilder) -> {
+			GenomeInterval txRegion = transcriptModelBuilder.getTXRegion();
+			if (txRegion == null) {
+				// Only warn if a transcript and not a gene, we only allow exons to be parts of genes as this is
+				// observed in RefSeq
+				LOGGER.debug("No transcript region for {}; skipping", parentId);
+			} else {
+				if (transcriptModelBuilder.getCDSRegion() == null) {
+					transcriptModelBuilder.setCDSRegion(new GenomeInterval(txRegion.getGenomeBeginPos(), 0));
+				}
+				transcriptModelsWithTxRegion.put(parentId, transcriptModelBuilder);
+			}
+		});
+		return transcriptModelsWithTxRegion;
+	}
+
+	private Map<String, TranscriptModelBuilder> mapNonRedundantTranscriptIdsToTranscriptModelBuilder(Map<String, TranscriptModelBuilder> transcriptModelsWithTxRegion, Map<String, List<String>> transciptIdToParentIds) {
+
+		Map<String, TranscriptModelBuilder> transcriptIdToTranscriptModelBuilders = new HashMap<>(transcriptModelsWithTxRegion
+			.size());
+
+		boolean assignPseudoAutosomalTranscriptsToX = preferPARTranscriptsOnChrX();
+		if (assignPseudoAutosomalTranscriptsToX) {
+			LOGGER.info("Pseudoautosomal transcripts will be assigned to X chromosome");
+		}
+
+		int duplicatedTranscriptIdCount = 0;
+		for (Entry<String, List<String>> entry : transciptIdToParentIds.entrySet()) {
+			String transcriptId = entry.getKey();
+			List<String> parentIds = entry.getValue();
+			if (parentIds.isEmpty()) {
+				continue;
+			}
+			if (parentIds.size() == 1) {
+				// transcriptModelsWithTxRegion used the ParentId as the key (something like 'rna58569'), but we need
+				// the final map to use the transcriptId e.g. 'XM_005255624.1' as the key
+				TranscriptModelBuilder transcriptModelBuilder = transcriptModelsWithTxRegion.get(parentIds.get(0));
+				transcriptIdToTranscriptModelBuilders.put(transcriptId, transcriptModelBuilder);
+			} else {
+				TranscriptModelBuilder existingTranscriptBuilder = transcriptModelsWithTxRegion.get(parentIds.get(0));
+				TranscriptModelBuilder updatedTranscriptBuilder = transcriptModelsWithTxRegion.get(parentIds.get(1));
+				int currentChromosome = existingTranscriptBuilder.getTXRegion().getChr();
+				int newChromosome = updatedTranscriptBuilder.getTXRegion().getChr();
+				// Pseudoautosomal gene
+				if (currentChromosome == 23 && newChromosome == 24) {
+					// e.g. https://www.ncbi.nlm.nih.gov/nuccore/NM_001636
+					TranscriptModelBuilder preferred = assignPseudoAutosomalTranscriptsToX ? existingTranscriptBuilder : updatedTranscriptBuilder;
+					LOGGER.info("Assigning pseudoautosomal gene {} transcript {} to chromsome {}",
+						preferred.getGeneSymbol(), preferred.getAccession(), preferred.getTXRegion().getChr());
+					transcriptIdToTranscriptModelBuilders.put(transcriptId, preferred);
+				} else {
+					duplicatedTranscriptIdCount++;
+					// This is a tricky call - there are about 30 transcripts with duplicated transcriptIds due to imperfect
+					// alignment to the genomic reference
+					LOGGER.warn("Transcript {} has {} possible transcript models - using the longest model", transcriptId, parentIds
+						.size());
+					List<TranscriptModelBuilder> possibleModels = parentIds.stream()
+						.map(transcriptModelsWithTxRegion::get)
+						.collect(toList());
+					TranscriptModelBuilder longestCds = findLongestTranscriptRegion(possibleModels);
+					transcriptIdToTranscriptModelBuilders.put(transcriptId, longestCds);
+				}
+			}
+		}
+		if (duplicatedTranscriptIdCount != 0) {
+			LOGGER.warn("{} duplicated transcript ids", duplicatedTranscriptIdCount);
+		}
+		return transcriptIdToTranscriptModelBuilders;
+	}
+
+	private TranscriptModelBuilder findLongestTranscriptRegion(List<TranscriptModelBuilder> possibleModels) {
+		TranscriptModelBuilder longestCds = possibleModels.get(0);
+		for (TranscriptModelBuilder current : possibleModels) {
+			if (current.getTXRegion().length() > longestCds.getTXRegion().length()) {
+				longestCds = current;
+			}
+		}
+		return longestCds;
+	}
 }