+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   Throwing new checked exception `InvalidGenomeVariant` case of error.
 * Fixing SO term ID for `VariantEffect.DISRUPTIVE_INFRAME_DELETION`
 * Correctly parsing transcript version for ENSEMBL when available (not available for b75/GRCh37).
+* Making transcript model building (for `download`) more memory efficient.
 
 ## v0.27